Session 2: Critical fixes and feature improvements
✅ Enhanced Landing Pages:
- Version 1 (BCG): Added architectural visualizations and blueprint elements
- Version 2 (Modern): Redesigned to light theme, removed emojis, added AI metrics

✅ Supplier Price Items Fixes:
- Added missing 'effective_from' column to supplier_price_items table
- Fixed price type from Integer to Float in model and schemas
- Updated CSV import to support both 'rate' and 'price' columns

✅ CSV Import for Supplier Price Lists:
- Fixed import logic to handle duplicate codes (update instead of error)
- Improved error handling with detailed messages
- Enhanced frontend dialog with format examples

✅ File Upload Fix:
- Fixed hardcoded 'demo-project' project_id in frontend
- Added auto-creation of default project on user registration
- Implemented dynamic project loading in FileUpload component
- Added Projects API (list, create, get, update, delete)

✅ Price Application Fix:
- Fixed AttributeError: changed 'element_type' to 'code' for matching
- Added missing unit_price and total_price columns to boq_items table
- Removed incorrect price conversion (pence to pounds)
- Fixed foreign key constraint error by not setting mapped_price_item_id

✅ Export Fix:
- Updated export logic to read unit_price from boq_items
- Implemented two-tier pricing system (supplier prices + admin price lists)
- Fixed CSV/XLSX/PDF exports to show correct prices

📝 Documentation:
- PROJECT_HISTORY.md: Complete session log with timestamps
- CSV_IMPORT_GUIDE.md: Detailed CSV import instructions
- QUICK_TEST_CSV_IMPORT.md: Quick testing guide
- SESSION_2_SUMMARY.md: Summary of all changes

Database migrations applied:
- ALTER TABLE supplier_price_items ADD COLUMN effective_from
- ALTER TABLE boq_items ADD COLUMN unit_price, total_price
- INSERT demo-project for current user

+++ backend/app/models/supplier.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Boolean, DateTime, ForeignKey, Integer
+from sqlalchemy import Column, String, Boolean, DateTime, ForeignKey, Integer, Float
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from uuid import uuid4
@@ -29,7 +29,7 @@ class SupplierPriceItem(Base):
     code = Column(String, nullable=False)
     description = Column(String, nullable=False)
     unit = Column(String, nullable=False)
-    price = Column(Integer, nullable=False)  # price in minor units (cents/pennies)
+    price = Column(Float, nullable=False)  # price as decimal (e.g., 10.50 for £10.50)
     currency = Column(String, default="GBP")
     effective_from = Column(DateTime(timezone=True), nullable=True)
     is_active = Column(Boolean, default=True)
